New API
The API is substantially different.

The new API uses players to decide what the next operation should be.
The Whist can hook into these players and tell them when it expects a
bet or a card placement, and they can do whatever they want with it.

There is also the direct player, which can use player models. A player
model is a logical algorithm which doesn't require user intervention and
can output a result "immediately"